+++ rts/motoko-rts/src/stabilization/deserialization/scan_stack.rs
