+++ rts/motoko-rts/src/stabilization/deserialization/stable_memory_access.rs
